requirements.txt
# FastAPI & WebSocket
fastapi==0.110.0
uvicorn==0.27.0
websockets>=11.0
python-dotenv==1.2.2

# Database (SQLite + ORM)
sqlalchemy==2.0.25

alembic==1.13.0

# Logging
structlog==24.1.0

# Data Validation
pydantic>=2.7.4
pydantic-settings>=2.10.1

# Existing dependencies (keep all - unchanged)
langchain==1.2.10
langchain-core==1.2.17
langchain-community==0.4.1
langchain-nvidia-ai-endpoints==1.1.0
langchain-text-splitters==1.1.1
chromadb==1.5.2
sentence-transformers==5.2.3
torch==2.10.0
transformers==5.2.0
numpy==2.2.6
pandas==2.3.3

# Gemini Live API
google-genai>=1.0.0